requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
animation==0.0.7
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
dhooks==1.1.4
discord.py==1.6.0
Django==2.2
emoji==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
flake8==3.8.4
future==0.18.2
idna==2.10
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.4
parso==0.7.1
pefile==2019.4.18
Pillow==8.2.0
playsound==1.2.2
plotly==4.14.3
pluggy==0.13.1
pluginbase==1.0.0
psutil==5.8.0
py-cpuinfo==7.0.0
pychalk==2.0.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.0.0
pyflakes==2.2.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pylint==2.8.2
PyNaCl==1.4.0
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PySide2==5.15.2
PySimpleGUI==4.11.0
PySimpleGUIQt==0.35.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytube==10.7.2
pytube3==9.6.4
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
retrying==1.3.3
rope==0.19.0
selenium==3.141.0
shiboken2==5.15.2
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
websocket-client==0.58.0
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
yfinance==0.1.59
youtube-dl==2021.4.1
